Src/CameraServerApp/requirements.txt
numpy==1.24.3
opencv-python-headless==4.7.0.72
pika==1.3.2
PyYAML==5.4.1